+++ apps/gitness/src/framework/routing/types.ts
@@ -1,0 +1,21 @@
+import { Params, RouteObject } from 'react-router-dom'
+
+// Enum defining the route constants
+export enum RouteConstants {
+  toRepoSummary = 'toRepoSummary',
+  toRepoCommits = 'toRepoCommits'
+}
+
+// Type for a mapping of enum keys to functions that generate paths
+export type RouteFunctionMap = Record<keyof typeof RouteConstants, (params: Params<string>) => string>
+
+// Define a custom handle with the breadcrumb property
+export interface CustomHandle {
+  breadcrumb?: (params: Params<string>) => string
+  routeName?: string
+}
+
+// Create a new type by intersecting RouteObject with the custom handle
+export type CustomRouteObject = RouteObject & {
+  handle?: CustomHandle
+}
